Sprint No.15.3.3 refactoring: change name package: ui/presentation remove class SearchHistory
