+++ app/src/main/java/com/example/playlistmaker/presentation/settings/App.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.ui.settings
+package com.example.playlistmaker.presentation.settings
 
 
 import android.app.Application
