+++ app/src/main/java/com/example/playlistmaker/presentation/track/TrackAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.ui.track
+package com.example.playlistmaker.presentation.track
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
