+++ app/src/main/res/layout/activity_audio_player.xml
@@ -11,7 +11,7 @@
         android:id="@+id/main"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.audio_player.AudioPlayer">
+        tools:context=".presentation.audio_player.AudioPlayer">
 
         <androidx.constraintlayout.widget.Guideline
             android:id="@+id/left_guideline_24"
