+++ app/src/main/res/layout/activity_search.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.search.SearchActivity">
+    tools:context=".presentation.search.SearchActivity">
 
     <!-- MaterialToolbar -->
     <com.google.android.material.appbar.MaterialToolbar
